+++ src/eca/config.clj
@@ -224,6 +224,7 @@
    :completion {:model "openai/gpt-4.1"}
    :netrcFile nil
    :autoCompactPercentage 75
+   :includeParentAgentsFiles false
    :plugins {"eca" {:source "https://github.com/editor-code-assistant/eca-plugins.git"}}
    :remote {:enabled false}
    :env "prod"})